+++ tests/test_convert.py
@@ -51,6 +51,7 @@ tests = [
     "ec_ro",
     "india_10k",
     "it_1",
+    "de_he",
 ]
 test_path = "tests/data-files/convert"
 
@@ -65,6 +66,7 @@ extra_convert_parameters = {
     "se": {"variant": "2023"},
     "si": {"variant": "2023"},
     "be_vlg": {"variant": "2023"},
+    "de_he": _input_files("de_he", "de_he.json"),
     "br_ba_lem": _input_files("br_ba_lem", "LEM_dataset.zip"),
     "ch": _input_files("ch", "lwb_nutzungsflaechen_v2_0_lv95.gpkg"),
     "es_cat": _input_files("es_cat", "Cultius_DUN2023_GPKG.zip"),