+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Update years for DK (2025, 2026)
 - Converter for Bavaria, Germany
 - Update fr-converter to support 2021/2022 files
+- Converter for Hesse, Germany LPIS reference parcels
 
 ## [v0.21.0] - 2026-02-16